services/model_based_event_detection/requirements.txt
paho-mqtt==1.6.1
rdflib==6.3.2